+++ test-forge/unit/tee/implementation/TeeInstructions.t.sol
@@ -21,7 +21,7 @@ contract TeeInstructionsTest is Test {
 
     event TeeInstructionsSent (
         bytes32 indexed instructionId,
-        uint24 indexed rewardEpochId,
+        uint32 indexed rewardEpochId,
         ITeeRegistry.TeeMachine[] teeMachines,
         bytes32 opType,
         bytes32 opCommand,
