+++ test-forge/unit/tee/implementation/TeeWalletManager.t.sol
@@ -41,7 +41,7 @@ contract TeeWalletManagerTest is Test {
 
     event TeeInstructionsSent(
         bytes32 indexed instructionId,
-        uint24 indexed rewardEpochId,
+        uint32 indexed rewardEpochId,
         ITeeRegistry.TeeMachine[] teeMachines,
         bytes32 opType,
         bytes32 opCommand,
